Zekin Personal Mentor/requirements.txt
requests
python-dotenv
openai
gradio
